Fitness.AI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pymongo==4.6.0
tensorflow==2.16.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
openai==1.3.5
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
requests==2.31.0 